Update example file name and use same setup as other examples
